requirements.txt
discord.py
schedule
python-dotenv
pyyaml